requirements.txt
# Core dependencies
numpy~=1.26.0
pandas~=2.1.0
scipy~=1.11.0
scikit-learn~=1.3.0
joblib~=1.3.0
threadpoolctl~=3.2.0

# Streamlit and extensions
streamlit~=1.41.0
streamlit-option-menu~=0.3.6
streamlit-javascript~=0.1.5
pydeck~=0.8.0

# Plotting and visualisation
altair~=5.2.0
pyarrow~=14.0.0
rich~=13.7.0
pillow~=10.2.0

# Utilities and helpers
attrs~=23.2.0
cachetools~=5.3.0
certifi~=2023.11.0
charset-normalizer~=3.3.0
click~=8.1.0
colorama~=0.4.6
idna~=3.6
importlib-metadata~=6.11.0
markdown-it-py~=3.0.0
MarkupSafe~=2.1.3
packaging~=23.2
protobuf~=4.25.0
python-dateutil~=2.8.0
pytz~=2023.3
requests~=2.31.0
setuptools~=68.0.0
six~=1.16.0
toml~=0.10.2
toolz~=0.12.0
typing_extensions~=4.9.0
tzdata~=2023.4
tzlocal~=5.2
urllib3~=2.1.0
validators~=0.22.0
watchdog~=3.0.0
wheel~=0.41.0
zipp~=3.17.0

# Advanced JSON handling
jsonschema~=4.20.0
jsonschema-specifications~=2023.12.1
referencing~=0.32.0

# Git integration
gitdb~=4.0.0
GitPython~=3.1.0
smmap~=5.0.0
